Simplify part upload
UploadPart class is now responsible for holding the state around the act
of upload the part but not the part itself. This is important because we
keep UploadPart instances for all uploaded parts so that we can complete
the upload at the end. But we definitely do not want to keep contents of
all parts until the end.